Refactor iterators.rs into a directory to fix The Blob anti-pattern
